requirements.txt
platypus-opt>=1.0.3
PyQt5>=5.14.2
PyQt5-stubs>=5.14.2.0